fixup review comments
