+++ kmod/patch/kpatch.lds
@@ -8,6 +8,13 @@ SECTIONS
     __kpatch_hooks_load = . ;
     *(.kpatch.hooks.load)
     __kpatch_hooks_load_end = . ;
+    /*
+     * Pad the end of the section with zeros in case the section is empty.
+     * This prevents the kernel from discarding the section at module
+     * load time. __kpatch_hooks_load_end will still point to the end of
+     * the section before the padding.  If the .kpatch.hooks.load section
+     * is empty, __kpatch_hooks_load equals __kpatch_hooks_load_end.
+     */
     QUAD(0);
   }
   .kpatch.hooks.unload : {
